+++ crawler/crawl_household.py
@@ -1,0 +1,6 @@
+# BaseCrawler 생성해 상속받기
+
+
+class HouseholdCrawler:
+    def __init__(self, building_code: str):
+        pass
