+++ crawler/object/household.py
@@ -1,0 +1,13 @@
+class Household:
+    def __init__(self, data: dict):
+        self.building_dong = None
+        self.floor = None
+        self.direction = None
+
+        self.supply_area = None
+        self.private_area = None
+
+        self.household_url = None
+        
+        self.fk_building_id = None
+
